+++ src/vcl/probe.rs
@@ -4,18 +4,52 @@ use std::time::Duration;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
-pub enum Request<'a> {
+pub enum COWRequest<'a> {
     URL(Cow<'a, str>),
     Text(Cow<'a, str>),
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct Probe<'a> {
-    pub request: Request<'a>,
+pub struct COWProbe<'a> {
+    pub request: COWRequest<'a>,
     pub timeout: Duration,
     pub interval: Duration,
     pub exp_status: c_uint,
     pub window: c_uint,
     pub threshold: c_uint,
     pub initial: c_uint,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub enum Request {
+    URL(String),
+    Text(String),
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct Probe {
+    pub request: Request,
+    pub timeout: Duration,
+    pub interval: Duration,
+    pub exp_status: c_uint,
+    pub window: c_uint,
+    pub threshold: c_uint,
+    pub initial: c_uint,
+}
+
+impl<'a> COWProbe<'a> {
+    pub fn to_owned(&self) -> Probe {
+        Probe {
+            request: match &self.request {
+                COWRequest::URL(cow) => Request::URL(cow.to_string()),
+                COWRequest::Text(cow) => Request::Text(cow.to_string()),
+            },
+            timeout: self.timeout,
+            interval: self.interval,
+            exp_status: self.exp_status,
+            window: self.window,
+            threshold: self.threshold,
+            initial: self.initial,
+        }
+    }
 }